requirements.txt
django
django-widget-tweaks
pillow
django-rest-framework
djangorestframework-datatables